backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-extensions==3.2.3
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
ortools>=9.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0